+++ MelonLoader/Libs/Harmony/PatchProcessor.cs
@@ -68,6 +68,8 @@ namespace Harmony
 					if (original == null)
 						throw new NullReferenceException("original");
 
+					if ((original.DeclaringType.Assembly.GetCustomAttributes(typeof(HarmonyShield), false).Count() > 0) || (original.GetCustomAttributes(typeof(HarmonyShield), false).Count() > 0)) continue;
+
 					var individualPrepareResult = RunMethod<HarmonyPrepare, bool>(true, original);
 					if (individualPrepareResult)
 					{